proximity: Add support for "proximity0" proximity sensors
This can be used in tsl2772 kernel driver, as well as devices with
multiple channels, such as combined proximity and light sensors, that
use "proximity0" as the channel name instead of "proximity".

This patch is inspired by 4dda433e920d3b4b8f286bf13246d2286aabab68,
although it is implemented slightly differently.