iio-buffer-utils: Correctly disable sensors when exiting
The enable_sensors() function has an error preventing sensor
disabling: it checks if the sensor is already enabled and does nothing
even if we need to disable it.

Also, there is lack of error handling in the _write_sysfs_int()
function: the result of fprintf() is not checked. If some error occurs
during writing to the sysfs attribute, it is ignored.

Fix these errors and change the order of sysfs attribute writing during
device releasing to avoid "Device or resource busy" errors (a buffer
should be disabled before scan elements).