tests: input_accel: Check for default value of AccelerometerOrientation
With the last commit we slightly changed the behavior of the input-accel
driver to be more in line with the other drivers: It now will only start
reading the sensor when ->set_polling(TRUE) was called. This means there's
no initial reading of the accelerometer available after calling ->open() on
the sensor.

That made the input_accel test fail, which now needs to check for the
default value of "AccelerometerOrientation", that is "undefined". After the
accelerometer has been claimed, the property should read "normal" again as
expected.