orientation: Report device tilt along z-axis
Expose information about the device being tilted or placed face up or down.
This property is potentially useful for implementing alarm or incoming call
silencing by placing the device face down, as done in Maemo and SailfishOS.

This is reported in a new AccelerometerTilt property.

Add tests for the new orientation states.