drivers: accel: Remove trigger requirement on iio-buffer-accel
IIO buffers can be used without having a trigger. Remove trigger requirement
on iio-buffer-accel to support devices that provide a buffer but no trigger
nor a poll interface.

This might break existing behavior relying on this hard requirement
as the fallback to the polling driver is not applied anymore for IIO
buffer sensors without triggers, see [1].

[1] https://gitlab.freedesktop.org/hadess/iio-sensor-proxy/-/merge_requests/368