iio-sensor-proxy: evaluate polling drivers ahead of buffering drivers
In order to support sensors on certain Qualcomm SoCs, commit 59c29aca61
("drivers: accel: Remove trigger requirement on iio-buffer-accel") drops
a test that checks for a valid trigger name before iio-buffer-accel
driver is bound on a device. Most other non-buffering accelerometers
depended on the failure of that test in order to fallback to the
iio-poll-accel driver. With the test gone, they are no longer recognized
by iio-sensor-proxy.

A proper fix would involve rethinking how iio-sensor-proxy should
differentiate between buffering and polling drivers.

In the meantime, a temporary fix is to rearrange the driver type
discovery so that poll drivers are tested first.

This will work fine for the Qualcomm SoCs in question, future Qualcomm
SSC sensors, and the majority of other sensors.

It will be a regression (from iio-buffer to iio-poll) for the very few
buffering drivers for which udev also provides the iio-poll property.
However, only devices with a trigger name that has the prefix
"accel_3d-dev" pass the `get_trigger_name()` test. It is reasonable to
assume that very few devices would have that prefix. In other words,
nearly everything was using the poll drivers in practice.

References:
 - https://gitlab.freedesktop.org/hadess/iio-sensor-proxy/-/issues/121
 - https://gitlab.freedesktop.org/hadess/iio-sensor-proxy/-/issues/332