buffer drivers: enable ring buffer only when sensor is claimed
Save a few kernel CPU cycles by disabling the buffer when the sensor is not
claimed. In order to do this, split disable_trigger off as a standalone
function from disable_ring_buffer, to allow disabling the ring buffer without
having to recreate the trigger information.

Adjust the buffer integration test accordingly. Previously, the buffer was
always enabled, so the test expects buffer/enabled = 1, which is no longer
always true.